move appdata to new module
